# spec/openapi.yaml
openapi: 3.1.0
info:
  title: QueueFlow API
  description: REST API for QueueFlow, a PostgreSQL/PGMQ-native distributed job queue and workflow engine.
  contact:
    name: QueueFlow
    url: https://queueflow.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: Local development
- url: https://api.queueflow.dev
  description: Production
paths:
  /api/v1/jobs:
    get:
      tags:
      - jobs
      operationId: listJobs
      parameters:
      - name: status
        in: query
        description: Filter by status (e.g. `pending`, `completed`).
        required: false
        schema:
          type: string
      - name: queue
        in: query
        description: Filter by queue name (jobs only).
        required: false
        schema:
          type: string
      - name: limit
        in: query
        description: Page size, 1..=100 (default 50).
        required: false
        schema:
          type: integer
          format: int64
      - name: offset
        in: query
        description: Number of records to skip (default 0).
        required: false
        schema:
          type: integer
          format: int64
      - name: order_by
        in: query
        description: '`created_at ASC` or `created_at DESC` (default DESC).'
        required: false
        schema:
          type: string
      - name: include_total
        in: query
        description: |-
          Include the exact `total` count in the response (default false; the
          count is an extra full scan over the filtered set).
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Page of jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListJobsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearerAuth: []
    post:
      tags:
      - jobs
      operationId: createJob
      parameters:
      - name: Idempotency-Key
        in: header
        description: 'Optional client-supplied key making this create idempotent per tenant: retrying with the same key returns the original job instead of creating a duplicate.'
        required: false
        schema:
          type:
          - string
          - 'null'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
        required: true
      responses:
        '201':
          description: Job created (or replayed idempotently)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateJobResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearerAuth: []
  /api/v1/jobs/batch:
    post:
      tags:
      - jobs
      operationId: createBatchJobs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBatchJobsRequest'
        required: true
      responses:
        '201':
          description: Jobs created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBatchJobsResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearerAuth: []
  /api/v1/jobs/{id}:
    get:
      tags:
      - jobs
      operationId: getJob
      parameters:
      - name: id
        in: path
        description: Job id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearerAuth: []
  /api/v1/jobs/{id}/cancel:
    post:
      tags:
      - jobs
      operationId: cancelJob
      parameters:
      - name: id
        in: path
        description: Job id
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Cancelled
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearerAuth: []
  /api/v1/jobs/{id}/complete:
    post:
      tags:
      - worker
      operationId: completeJob
      parameters:
      - name: id
        in: path
        description: Job id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteJobRequest'
        required: true
      responses:
        '204':
          description: 'Completed (idempotent: replaying a finished lease also succeeds)'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearerAuth: []
  /api/v1/jobs/{id}/events:
    get:
      tags:
      - jobs
      summary: |-
        Stream a job's status transitions as Server-Sent Events until it reaches a
        terminal state. Lets clients await completion without polling the REST
        endpoint themselves.
      operationId: streamJobEvents
      parameters:
      - name: id
        in: path
        description: Job id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: SSE stream; each `status` event carries the full job JSON. Closes after the job reaches a terminal state.
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearerAuth: []
  /api/v1/jobs/{id}/fail:
    post:
      tags:
      - worker
      operationId: failJob
      parameters:
      - name: id
        in: path
        description: Job id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FailJobRequest'
        required: true
      responses:
        '204':
          description: Failure recorded; the job is retried or dead-lettered per its config
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearerAuth: []
  /api/v1/jobs/{id}/heartbeat:
    post:
      tags:
      - worker
      operationId: heartbeatJob
      parameters:
      - name: id
        in: path
        description: Job id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeartbeatRequest'
        required: true
      responses:
        '204':
          description: Lease extended
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearerAuth: []
  /api/v1/queues/{queue}/lease:
    post:
      tags:
      - worker
      operationId: leaseJobs
      parameters:
      - name: queue
        in: path
        description: Queue to lease from
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaseJobsRequest'
        required: true
      responses:
        '200':
          description: Zero or more leased jobs (empty if none became available within wait_secs)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaseJobsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearerAuth: []
  /api/v1/stats:
    get:
      tags:
      - system
      operationId: getStats
      responses:
        '200':
          description: 'Engine counters. Process-local and reset on restart: in a split api/worker deployment this reflects only the process serving the request; query the database for fleet-wide history.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsSnapshot'
      security:
      - bearerAuth: []
  /api/v1/tasks:
    get:
      tags:
      - system
      operationId: listTasks
      responses:
        '200':
          description: Registered task handlers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TasksResponse'
      security:
      - bearerAuth: []
...
  /api/v1/workflows/{id}:
    get:
      tags:
      - workflows
      operationId: getWorkflow
      parameters:
      - name: id
        in: path
        description: Workflow id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Workflow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearerAuth: []
  /api/v1/workflows/{id}/cancel:
    post:
      tags:
      - workflows
      operationId: cancelWorkflow
      parameters:
      - name: id
        in: path
        description: Workflow id
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Cancelled
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearerAuth: []
  /api/v1/workflows/{id}/diagram:
    get:
      tags:
      - workflows
      operationId: getWorkflowDiagram
      parameters:
      - name: id
        in: path
        description: Workflow id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Mermaid diagram of the workflow DAG
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDiagramResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    CreateJobRequest:
      type: object
      required:
      - task_name
      properties:
        config:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/JobConfigRequest'
        payload:
          type: object
          description: Arbitrary JSON object passed to the handler.
        task_name:
          type: string
          description: The registered task handler to invoke.
...
    ErrorBody:
      type: object
      required:
      - error
      - timestamp
      properties:
        error:
          type: string
        timestamp:
          type: string
          format: date-time
...
    ListJobsResponse:
      type: object
      required:
      - jobs
      - limit
      - offset
      - has_more
      properties:
        has_more:
          type: boolean
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/Job'
        limit:
          type: integer
          format: int64
        offset:
          type: integer
          format: int64
        total:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Exact total match count. Only present when the request set
            `include_total=true`; computing it costs a full count over the filtered
            set, so it is opt-in.
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
...